Fix modded oxidizable blocks not oxidizing
Fixes #4371

`OxidizableBlocksRegistry.registerOxidizableBlockPair` will now re-cache the value of the `ticksRandomly` field of the passed blocks' states to ensure that the default implementation of oxidization will actually work on the blocks.

This may be considered a slightly dirty workaround, but it works well and isn't particularly intrusive, despite using 2 mixins.

I've also included a test (though not a GameTest) to ensure that this works, and split the `OxidizableBlocksRegistry` into separate API and impl classes to more easily allow for potential extensibility in future while not breaking compatibility.