Fix behavior change in `SingleVariantStorage` serialization
25w09a port added `orElseThrow()` call in `readNbt`, changing the behavior for
missing `variant` field. This previously used `fallback` (or, if the `codec` was
somehow a unit codec/could accept empty NBT, successfully decoded), but the change
made it throw. This makes it use the fallback in both cases.
Given that it is unlikely for a variant to use a unit codec, the remaining change is
deemed acceptable.