Fix getRecipeIdentifier applying twice to advancement paths
Fixes a bug introduced in #4703.

The advancement's path is already modified by the getRecipeIdentifier call
in each recipe JSON builder mixin, so this applied getRecipeIdentifier twice
to a single ID (and to different parts).

For example, this getRecipeIdentifier implementation:
    return super.getRecipeIdentifier(identifier).withPrefixedPath("foo/");
would have resulted in an advancement being placed at
    data/mod_id/advancements/foo/recipes/.../foo/the_item.json,
which is clearly wrong. With this fix, only the last "foo/" will appear.