Use RegistryInfoGetter to test TagsPopulatedResourceCondition
This should work based on how PendingTags are implemented and how the WrapperLookup is constructed during a server resource reload. Tags retrieved through the RegistryInfoGetter will be present as expected, but their contents will be empty unless the registry is a reloadable registry (loot tables). Luckily this resource condition does not need to test for tag contents.

This also allows the condition to work even after the resource reload is done.